Chore: rename Booklist css & move Booklist.js to components
