+++ package.json
@@ -11,6 +11,7 @@
         "react-dom": "^18.2.0",
         "react-router-dom": "^6.21.3",
         "react-scripts": "5.0.1",
+        "recoil": "^0.7.7",
         "web-vitals": "^2.1.0"
     },
     "scripts": {
